requirements.txt
asgiref==3.8.1
snowflake-id==1.0.2
dj-database-url==2.3.0
Django>=4.1,<5.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2
django-environ==0.11.2